#45 Isolated AddUser method to new service
