+++ code/FinanceManager.Domain/Services/IUserService.cs
@@ -1,0 +1,7 @@
+﻿namespace FinanceManager.Domain.Services
+{
+    public interface IUserService
+    {
+        Task<bool> AddUser(string login, string password);
+    }
+}
